port hobbes to OSX (#5)
get hobbes to build and pass tests on OSX 10.12.5 w/ LLVM 3.5 and 3.7

(just the class conversion tests fail because clang doesn't give the same RTTI structure as gcc [AFAIK], necessary to correctly convert)

+++ lib/hobbes/eval/funcdefs.C
@@ -457,7 +457,7 @@ void* runClosThread(void* clos) {
 long newThread(char* clos) {
   pthread_t th;
   pthread_create(&th, 0, &runClosThread, (void*)clos);
-  return th;
+  return (long)th;
 }
 
 void threadWait(long x) {
